# facturation-planning-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API pour gérer la facturation et les plannings",
        "title": "Facturation API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/clients": {
            "get": {
                "description": "Retourne la liste complète de tous les clients",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "Récupérer tous les clients",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Client"
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Crée un nouveau client avec les informations fournies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "Créer un nouveau client",
                "parameters": [
                    {
                        "description": "Informations du client",
                        "name": "client",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Client"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Client"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/login": {
            "post": {
                "description": "Permet de se connecter avec un email et un mot de passe pour récupérer un token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion d'une entreprise",
                "parameters": [
                    {
                        "description": "Identifiants de connexion",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token JWT généré",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Email ou mot de passe incorrect",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/register": {
            "post": {
                "description": "Permet de créer un compte entreprise en renseignant un nom, un email et un mot de passe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentification"
                ],
                "summary": "Créer une entreprise (inscription)",
                "parameters": [
                    {
                        "description": "Détails de l'entreprise",
                        "name": "entreprise",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterEntrepriseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Email déjà utilisé",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retourne les informations de l'entreprise actuellement connectée",
                "produces": [
                    "application/json"
                ],
                "summary": "Récupérer le profil de l'entreprise connectée",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Entreprise"
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Client": {
            "type": "object",
            "properties": {
                "adresse": {
                    "type": "string",
                    "example": "123 rue de la Paix, 75001 Paris"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-06-10T10:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "contact@acme.com"
                },
                "entreprise_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "nom": {
                    "type": "string",
                    "example": "ACME Corporation"
                },
                "telephone": {
                    "type": "string",
                    "example": "+33123456789"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-06-10T10:00:00Z"
                }
            }
        },
...
        "models.Entreprise": {
            "type": "object",
            "properties": {
                "adresse": {
                    "type": "string",
                    "example": "123 rue du Code"
                },
                "bic": {
                    "type": "string",
                    "example": "BIC12345"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-03-17T14:09:30.706109+01:00"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "example": "contact@maboite.com"
                },
                "iban": {
                    "type": "string",
                    "example": "FR7612345987650123456789014"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "nom": {
                    "type": "string",
                    "example": "MaBoite"
                },
                "responsable": {
                    "type": "string",
                    "example": "Jean Dupont"
                },
                "siret": {
                    "type": "string",
                    "example": "12345678901234"
                },
                "telephone": {
                    "type": "string",
                    "example": "+33123456789"
                },
                "tva": {
                    "type": "string",
                    "example": "FR12345678901"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-03-17T14:09:30.706109+01:00"
                }
            }
        },
...
        "models.LoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "contact@maboite.com"
                },
                "password": {
                    "type": "string",
                    "example": "securepass"
                }
            }
        },
...
        "models.RegisterEntrepriseRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "contact@maboite.com"
                },
                "nom": {
                    "type": "string",
                    "example": "MaBoite"
                },
                "password": {
                    "type": "string",
                    "example": "securepass"
                }
            }
        },
...
    }
}